Updating user table

+++ src/main/java/com/rs2/notetaking/controller/NoteTakingController.java
@@ -48,7 +48,7 @@ public class NoteTakingController {
     }
 
     @DeleteMapping(path = "/{id}")
-    public void deleteNote(@PathVariable(value = "id") int id) {
+    public void deleteNote(@PathVariable(value = "id") Long id) {
         noteService.deleteNote(id);
     }
