+++ src/main/java/com/rs2/notetaking/dto/NoteUpdateDTO.java
@@ -8,13 +8,13 @@ import lombok.Data;
 @Data
 public class NoteUpdateDTO {
 
-    private int noteId;
+    private Long noteId;
     private String title;
     private String content;
     private List<String> labels = new ArrayList<>();
 
 
-    public NoteUpdateDTO(int noteId, String title, String content, List<String> labels) {
+    public NoteUpdateDTO(Long noteId, String title, String content, List<String> labels) {
         this.noteId = noteId;
         this.title = title;
         this.content = content;
