+++ src/main/java/com/rs2/notetaking/service/NoteTakingService.java
@@ -17,7 +17,7 @@ public interface NoteTakingService {
  
     Optional<NoteDetailsDTO> updateNote(NoteUpdateDTO noteUpdateDto);
  
-    boolean deleteNote(int id);
+    boolean deleteNote(Long id);
 
     // Ideally implement this in the Label Service
     List<Label> getAllLabels();
